+++ src/components/common/atoms/Buttons.jsx
@@ -12,16 +12,24 @@ type Props = {
 
 const defaultProps = {
   id: undefined,
-  className: undefined
+  className: undefined,
+  disabled: false
 };
 
-export const PrimaryButton = ({ id, onClick, children, className , disabled}: Props) => (
+export const PrimaryButton = ({
+  id,
+  onClick,
+  children,
+  className,
+  disabled
+}: Props) => (
   <button
     id={id}
     disabled={disabled || false}
     type="button"
     className={classnames('primary-button', className)}
     onClick={onClick}
+    disabled={disabled}
   >
     {children}
   </button>