+++ src/redux/reducers/proposal.js
@@ -34,7 +34,10 @@ const {
   ON_QUESTIONS_FILTERED,
   RESET_QUESTIONS_FILTER,
   CLEAR_QUESTIONS_FILTER,
-  EXPAND_ALL_SECTIONS
+  EXPAND_ALL_SECTIONS,
+  SET_EDIT_QUESTION_DATA,
+  PROPOSAL_EDIT_QUESTION,
+  PROPOSAL_DELETE_QUESTION
 } = REDUX_TYPES.PROPOSAL;
 
 const INITIAL_STATE: Map = fromJS({
@@ -76,7 +79,8 @@ const INITIAL_STATE: Map = fromJS({
     }
   }),
   filteredProposalQuestions: Map({}),
-  areAllSectionsExpanded: false
+  areAllSectionsExpanded: false,
+  editQuestionsData: Map({})
 });
 
 const onProsalInfoLoaded = (state: Map, action: Object): Map => {
@@ -326,6 +330,48 @@ const onExpandAllSections = (state, action) => {
   return state.set('areAllSectionsExpanded', action.payload);
 };
 
+const onSetEditQuestionData = (state, action) => {
+  return state.set('editQuestionsData', fromJS(action.payload));
+};
+
+const onEditQuestion = (state, action) => {
+  const { payload: data } = action;
+  const questions = state.get('proposalQuestions');
+  const questionIndex = questions.findIndex(
+    item => item.questionId === data.questionId
+  );
+
+  const updatedQuestions = [
+    ...questions.slice(0, questionIndex),
+    data,
+    ...questions.slice(questionIndex + 1, questions.length)
+  ];
+
+  return state
+    .set('proposalQuestions', cloneDeep(updatedQuestions))
+    .set('setQuestionData', data)
+    .set('isSetQuestionLoading', false);
+};
+
+const onDeleteQuestion = (state, action) => {
+  const { payload: questionId } = action;
+  const questions = state.get('proposalQuestions');
+
+  const questionIndex = questions.findIndex(
+    item => item.questionId === questionId
+  );
+
+  const updatedQuestions = [
+    ...questions.slice(0, questionIndex),
+    ...questions.slice(questionIndex + 1, questions.length)
+  ];
+
+  return state
+    .set('proposalQuestions', cloneDeep(updatedQuestions))
+    .set('setQuestionData', questionId)
+    .set('isSetQuestionLoading', false);
+};
+
 const actionMap = {
   [PROPOSAL_INFO]: onProsalInfoLoaded,
   [PROPOSAL_INFO_LOADING]: onProposalLoading,
@@ -356,7 +402,10 @@ const actionMap = {
   [ON_QUESTIONS_FILTERED]: onQuestionsFiltered,
   [RESET_QUESTIONS_FILTER]: resetQuestionsFilter,
   [CLEAR_QUESTIONS_FILTER]: clearQuestionsFilter,
-  [EXPAND_ALL_SECTIONS]: onExpandAllSections
+  [EXPAND_ALL_SECTIONS]: onExpandAllSections,
+  [SET_EDIT_QUESTION_DATA]: onSetEditQuestionData,
+  [PROPOSAL_EDIT_QUESTION]: onEditQuestion,
+  [PROPOSAL_DELETE_QUESTION]: onDeleteQuestion
 };
 
 export default function(